Updates pause screen

+++ BulletHell.cpp
@@ -6,7 +6,7 @@
 #include <ctime>
 
 const int ScreenHeight = 1280;
-const int ScreenWidth = ScreenHeight * 0.75; // 960px -> Emulate a 4:3 portrait monitor as the ones used in old shoot'em ups
+const int ScreenWidth = (int)(ScreenHeight * 0.75); // 960px -> Emulate a 4:3 portrait monitor as the ones used in old shoot'em ups
 
 int main(void) {
